+++ src/Bootstrap.php
@@ -5,12 +5,14 @@ namespace davidhirtz\yii2\cms\hotspot;
 use davidhirtz\yii2\cms\hotspot\models\events\AssetAfterDuplicateEventHandler;
 use davidhirtz\yii2\cms\hotspot\models\events\AssetBeforeDeleteEventHandler;
 use davidhirtz\yii2\cms\hotspot\models\events\AssetBeforeDuplicateEventHandler;
+use davidhirtz\yii2\cms\hotspot\models\events\FileBeforeDeleteEventHandler;
 use davidhirtz\yii2\cms\hotspot\models\HotspotAsset;
 use davidhirtz\yii2\cms\hotspot\modules\admin\Module;
 use davidhirtz\yii2\cms\models\Asset;
 use davidhirtz\yii2\cms\models\builders\EntrySiteRelationsBuilder;
 use davidhirtz\yii2\cms\modules\admin\widgets\grids\columns\AssetThumbnailColumn;
 use davidhirtz\yii2\cms\widgets\Canvas;
+use davidhirtz\yii2\media\models\File;
 use davidhirtz\yii2\media\modules\admin\widgets\forms\fields\AssetPreview;
 use davidhirtz\yii2\skeleton\models\actions\DuplicateActiveRecord;
 use davidhirtz\yii2\skeleton\models\events\DuplicateActiveRecordEvent;
@@ -47,9 +49,7 @@ class Bootstrap implements BootstrapInterface
                 ],
             ],
             'media' => [
-                'assets' => [
-                    HotspotAsset::class,
-                ],
+                'fileRelations' => [HotspotAsset::class],
             ],
         ]);
 
@@ -69,19 +69,39 @@ class Bootstrap implements BootstrapInterface
         ModelEvent::on(
             Asset::class,
             Asset::EVENT_BEFORE_DELETE,
-            fn (ModelEvent $event) => Yii::createObject(AssetBeforeDeleteEventHandler::class, [$event])
+            fn (ModelEvent $event) => Yii::createObject(AssetBeforeDeleteEventHandler::class, [
+                $event,
+                $event->sender,
+            ])
         );
 
         ModelEvent::on(
             Asset::class,
             DuplicateActiveRecord::EVENT_BEFORE_DUPLICATE,
-            fn (DuplicateActiveRecordEvent $event) => Yii::createObject(AssetBeforeDuplicateEventHandler::class, [$event])
+            fn (DuplicateActiveRecordEvent $event) => Yii::createObject(AssetBeforeDuplicateEventHandler::class, [
+                $event,
+                $event->sender,
+                $event->duplicate,
+            ])
         );
 
         ModelEvent::on(
             Asset::class,
             DuplicateActiveRecord::EVENT_AFTER_DUPLICATE,
-            fn (DuplicateActiveRecordEvent $event) => Yii::createObject(AssetAfterDuplicateEventHandler::class, [$event])
+            fn (DuplicateActiveRecordEvent $event) => Yii::createObject(AssetAfterDuplicateEventHandler::class, [
+                $event,
+                $event->sender,
+                $event->duplicate,
+            ])
+        );
+
+        ModelEvent::on(
+            File::class,
+            File::EVENT_BEFORE_DELETE,
+            fn (ModelEvent $event) => Yii::createObject(FileBeforeDeleteEventHandler::class, [
+                $event,
+                $event->sender,
+            ])
         );
 
         $app->setMigrationNamespace('davidhirtz\yii2\cms\hotspot\migrations');