+++ src/models/events/AssetAfterDuplicateEventHandler.php
@@ -13,25 +13,25 @@ use Yii;
  */
 class AssetAfterDuplicateEventHandler
 {
-    public function __construct(protected DuplicateActiveRecordEvent $event)
-    {
+    public function __construct(
+        protected readonly DuplicateActiveRecordEvent $event,
+        protected readonly Asset $asset,
+        protected readonly Asset $duplicate,
+    ) {
         $this->handleEvent();
     }
 
     public function handleEvent(): void
     {
-        /** @var Asset $asset */
-        $asset = $this->event->sender;
-
-        if ($asset->getAttribute('hotspot_count')) {
+        if ($this->asset->getAttribute('hotspot_count')) {
             Yii::debug('Duplicating hotspots ...');
 
-            $hotspots = Hotspot::findAll(['asset_id' => $asset->id]);
+            $hotspots = Hotspot::findAll(['asset_id' => $this->asset->id]);
 
             foreach ($hotspots as $hotspot) {
                 DuplicateHotspot::create([
                     'hotspot' => $hotspot,
-                    'asset' => $this->event->duplicate,
+                    'asset' => $this->duplicate,
                     'shouldUpdateAssetAfterInsert' => false,
                     'attributes' => [
                         'status' => $hotspot->status,